move submitHashes into own file
