+++ lib/utils/helpers.js
@@ -1,7 +1,7 @@
 import fs from 'fs'
 import crypto from 'crypto'
 import uuidValidate from 'uuid-validate'
-import { isEmpty, has, isObject } from 'lodash'
+import { isEmpty, has, isObject, isArray, reject } from 'lodash'
 
 /**
  * Checks if value is a hexadecimal string
@@ -104,4 +104,36 @@ export function promiseMap(arr, fn) {
       }
     })
   )
+}
+
+/**
+ * Get SHA256 hash(es) of selected file(s) and prepare for submitting to a node
+ * @param {Array<String>} paths - An Array of paths of the files to be hashed.
+ * @param {Array<String>} uris - An Array of String URI's. Each hash will be submitted to each Node URI provided. If none provided three will be chosen at random using service discovery.
+ * @return {Array<{path: String, hash: String} An Array of Objects, each a handle that contains all info needed to retrieve a proof.
+ */
+export async function getFileHashes(paths, uris) {
+  uris = uris || []
+
+  // Validate all paths provided
+  if (!isArray(paths)) throw new Error('paths arg must be an Array')
+  if (isEmpty(paths)) throw new Error('paths arg must be a non-empty Array')
+  if (paths.length > 250) throw new Error('paths arg must be an Array with <= 250 elements')
+  let rejects = reject(paths, path => fs.existsSync(path) && fs.lstatSync(path).isFile())
+  if (!isEmpty(rejects)) throw new Error(`paths arg contains invalid paths : ${rejects.join(', ')}`)
+
+  // Validate all Node URIs provided
+  if (!isArray(uris)) throw new Error('uris arg must be an Array of String URIs')
+  if (uris.length > 5) throw new Error('uris arg must be an Array with <= 5 elements')
+
+  let hashObjs = []
+  hashObjs = await Promise.all(paths.map(path => sha256FileByPath(path)))
+
+  // filter out any EACCES errors
+  hashObjs = hashObjs.filter(hashObj => {
+    if (hashObj.error === 'EACCES') console.error(`Insufficient permission to read file '${hashObj.path}', skipping`)
+    return hashObj.error !== 'EACCES'
+  })
+  console.log('hashObjs:', hashObjs)
+  return hashObjs
 }
