+++ lib/utils/index.js
@@ -17,4 +17,5 @@ import * as helpers from './helpers'
 import * as proofs from './proofs'
 import * as network from './network'
 
+export { helpers, proofs, network }
 export default { ...helpers, ...proofs, ...network }
