+++ lib/utils/network.js
@@ -11,6 +11,11 @@
  * limitations under the License.
  */
 
+/*
+ * helper functions related to interact with chainpoint network objects
+ * such as nodes and cores
+ */
+
 const dns = require('dns')
 import url from 'url'
 import { isInteger, isFunction, isEmpty, slice, map, shuffle, filter, first, isString } from 'lodash'
